+++ packages/common/src/utils/signature_utils.ts
@@ -5,10 +5,15 @@ import { getEIP191Verifier } from '@didtools/pkh-ethereum'
 import { getSolanaVerifier } from '@didtools/pkh-solana'
 import { getStacksVerifier } from '@didtools/pkh-stacks'
 import { getTezosVerifier } from '@didtools/pkh-tezos'
+import { WebauthnAuth } from '@didtools/key-webauthn'
 import { CeramicSigner } from '../ceramic-signer.js'
 import { StreamUtils } from './stream-utils.js'
+import { ServiceMetrics as Metrics } from '@ceramicnetwork/observability'
+import { ModelMetrics } from '@ceramicnetwork/model-metrics'
 
 const DEFAULT_CACAO_REVOCATION_PHASE_OUT_SECS = 24 * 60 * 60
+
+export const CACAO_EXPIRED = 'cacao_expired'
 
 // Register supported CACAO Verifiers
 const verifiersCACAO = {
@@ -16,6 +21,7 @@ const verifiersCACAO = {
   ...getSolanaVerifier(),
   ...getStacksVerifier(),
   ...getTezosVerifier(),
+  ...WebauthnAuth.getVerifier(),
 }
 
 /**
@@ -52,6 +58,11 @@ export class SignatureUtils {
       })
     } catch (e: any) {
       const original = e.message ? e.message : String(e)
+      if (original.includes('CACAO has expired')) {
+        // TODO: string matching error messages is brittle. Can we use a stable error code instead?
+        Metrics.count(CACAO_EXPIRED, 1, { source: 'new_commit' })
+        ModelMetrics.recordError(CACAO_EXPIRED + '_new_commit')
+      }
       throw new Error(
         `Can not verify signature for commit ${commitData.cid} to stream ${streamId} which has controller DID ${controller}: ${original}`
       )
@@ -101,6 +112,8 @@ export class SignatureUtils {
       }
       const expirationTime = logEntry.expirationTime + DEFAULT_CACAO_REVOCATION_PHASE_OUT_SECS
       if (expirationTime < timestamp) {
+        Metrics.count(CACAO_EXPIRED, 1, { source: 'existing_state' })
+        ModelMetrics.recordError(CACAO_EXPIRED + '_existing_state')
         throw new Error(
           `CACAO expired: Commit ${logEntry.cid.toString()} of Stream ${StreamUtils.streamIdFromState(
             state