+++ packages/core/src/store/anchor-request-store.ts
@@ -114,7 +114,7 @@ export class AnchorRequestStore extends ObjectStore<StreamID, AnchorRequestData>
             resolve(null)
           }, this.#infiniteListBatchTimeoutMs)
         })
-        this.#logger.debug(`Fetching batch from AnchorRequestStore starting at key ${gt}`)
+        this.#logger.verbose(`Fetching batch from AnchorRequestStore starting at key ${gt}`)
         const batchPromise = this.store.find({
           limit: batchSize,
           gt: generateKey(gt),